pyproject.toml
[project]
name = "dunnobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
